requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-gis==1.0
fiona==1.9.5
geopandas==0.13.2
idna==3.4
importlib-metadata==6.8.0
joblib==1.3.2
networkx==3.1
numpy==1.24.4
osmnx==1.7.1
packaging==23.2
pandas==2.0.3
psycopg2-binary==2.9.9
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
zipp==3.17.0
